+++ taipei-day-trip/app/app.py
@@ -1,0 +1,77 @@
+# The line sys.stdout.reconfigure(encoding='utf-8') attempts to reconfigure the standard output stream(sys.stdout) to use UTF-8 encoding
+from fastapi.staticfiles import StaticFiles
+from fastapi.responses import FileResponse, JSONResponse
+from fastapi import *
+from fastapi.encoders import jsonable_encoder
+from fastapi.exceptions import RequestValidationError
+from starlette.exceptions import HTTPException as StarletteHTTPException
+
+import sys
+sys.stdout.reconfigure(encoding='utf-8')
+
+# router
+
+from .routers import attractions, user
+
+# Server
+
+app = FastAPI()
+
+# Router include
+
+app.include_router(attractions.router)
+app.include_router(user.router)
+
+# Static files
+
+app.mount("/style", StaticFiles(directory="style"), name="style")
+app.mount("/source", StaticFiles(directory='source'), name='source')
+app.mount("/function", StaticFiles(directory="function"), name='function')
+
+
+# Static Pages (Never Modify Code in this Block)
+
+
+@app.get("/", include_in_schema=False)
+async def index(request: Request):
+    return FileResponse("./static/index.html", media_type="text/html")
+
+
+@app.get("/attraction/{id}", include_in_schema=False)
+async def attraction(request: Request, id: int):
+    return FileResponse("./static/attraction.html", media_type="text/html")
+
+
+@app.get("/booking", include_in_schema=False)
+async def booking(request: Request):
+    return FileResponse("./static/booking.html", media_type="text/html")
+
+
+@app.get("/thankyou", include_in_schema=False)
+async def thankyou(request: Request):
+    return FileResponse("./static/thankyou.html", media_type="text/html")
+
+
+# Error handler
+
+
+@app.exception_handler(RequestValidationError)
+async def validation_exception_handler(request: Request, exc: RequestValidationError):
+    return JSONResponse(
+        status_code=400,
+        content=jsonable_encoder({
+            "error": True,
+            "message": str(f"{exc.errors()[0]['loc'][1]}:{exc.errors()[0]['msg']}")
+        })
+    )
+
+
+@app.exception_handler(StarletteHTTPException)
+async def http_exception_handler(request, exc):
+    return JSONResponse(
+        status_code=exc.status_code,
+        content=jsonable_encoder({
+            "error": True,
+            "message": str(exc.detail)
+        })
+    )
